refactor search index to use "CurrentRun"
