+++ runtime-common/realm.ts
@@ -93,7 +93,6 @@ export class Realm {
     this.#startedUp.fulfill((() => this.#startup())());
     this.#searchIndex = new SearchIndex(
       this,
-      this.paths,
       this.#adapter.readdir.bind(this.#adapter),
       this.readFileAsText.bind(this)
     );
