+++ runtime-common/search-index.ts
@@ -1,18 +1,9 @@
 import { executableExtensions, baseRealm } from ".";
-import {
-  Kind,
-  Realm,
-  CardDefinitionResource,
-  getExportedCardContext,
-} from "./realm";
-import { RealmPaths, LocalPath } from "./paths";
-import { ModuleSyntax } from "./module-syntax";
-import { ClassReference, PossibleCardClass } from "./schema-analysis-plugin";
-import ignore, { Ignore } from "ignore";
-import { stringify } from "qs";
+import { Kind, Realm } from "./realm";
+import { CurrentRun } from "./current-run";
+import { LocalPath } from "./paths";
 import { Query, Filter, Sort } from "./query";
 import { Loader } from "./loader";
-import { Deferred } from "./deferred";
 //@ts-ignore realm server TSC doesn't know how to deal with this because it doesn't understand glint
 import type { Card } from "https://cardstack.com/base/card-api";
 //@ts-ignore realm server TSC doesn't know how to deal with this because it doesn't understand glint
@@ -139,7 +130,7 @@ export interface CardDefinition {
   >;
 }
 
-function hasExecutableExtension(path: string): boolean {
+export function hasExecutableExtension(path: string): boolean {
   for (let extension of executableExtensions) {
     if (path.endsWith(extension)) {
       return true;
@@ -148,57 +139,13 @@ function hasExecutableExtension(path: string): boolean {
   return false;
 }
 
-function trimExecutableExtension(url: URL): URL {
+export function trimExecutableExtension(url: URL): URL {
   for (let extension of executableExtensions) {
     if (url.href.endsWith(extension)) {
       return new URL(url.href.replace(new RegExp(`\\${extension}$`), ""));
     }
   }
   return url;
-}
-
-// Forces callers to use URL (which avoids accidentally using relative url
-// strings without a base)
-class URLMap<T> {
-  #map: Map<string, T>;
-  constructor(mapTuple: [key: URL, value: T][] = []) {
-    this.#map = new Map(mapTuple.map(([key, value]) => [key.href, value]));
-  }
-  get(url: URL): T | undefined {
-    return this.#map.get(url.href);
-  }
-  set(url: URL, value: T) {
-    return this.#map.set(url.href, value);
-  }
-  get [Symbol.iterator]() {
-    let self = this;
-    return function* () {
-      for (let [key, value] of self.#map) {
-        yield [new URL(key), value] as [URL, T];
-      }
-    };
-  }
-  values() {
-    return this.#map.values();
-  }
-  keys() {
-    let self = this;
-    return {
-      get [Symbol.iterator]() {
-        return function* () {
-          for (let key of self.#map.keys()) {
-            yield new URL(key);
-          }
-        };
-      },
-    };
-  }
-  get size() {
-    return this.#map.size;
-  }
-  remove(url: URL) {
-    return this.#map.delete(url.href);
-  }
 }
 
 interface SearchEntry {
@@ -208,344 +155,38 @@ interface SearchEntry {
 }
 
 export class SearchIndex {
-  private instances = new URLMap<SearchEntry>();
-  private modules = new URLMap<ModuleSyntax>();
-  private definitions = new Map<string, CardDefinition>();
-  private exportedCardRefs = new Map<string, Map<string, ExportedCardRef>>();
-  private ignoreMap = new URLMap<Ignore>();
-  #api: CardAPI | undefined;
-  #externalDefinitionsCache = new Map<
-    string,
-    Promise<CardDefinition | undefined>
-  >();
+  #currentRun: CurrentRun;
 
   constructor(
     private realm: Realm,
-    private realmPaths: RealmPaths,
     private readdir: (
       path: string
     ) => AsyncGenerator<{ name: string; path: string; kind: Kind }, void>,
     private readFileAsText: (
       path: LocalPath
     ) => Promise<{ content: string; lastModified: number } | undefined>
-  ) {}
+  ) {
+    this.#currentRun = CurrentRun.empty(realm);
+  }
+
+  // TODO this is for debugging only--delete this
+  public get currentRun() {
+    return this.#currentRun;
+  }
 
   async run() {
-    this.#api = await Loader.import<CardAPI>(`${baseRealm.url}card-api`);
-    await this.visitDirectory(new URL(this.realm.url));
-    await this.semanticPhase();
-  }
-
-  private get api(): CardAPI {
-    if (!this.#api) {
-      throw new Error(`Card API was accessed before it was loaded`);
-    }
-    return this.#api;
-  }
-
-  private async visitDirectory(url: URL): Promise<void> {
-    let ignorePatterns = await this.readFileAsText(
-      this.realmPaths.local(new URL(".gitignore", url))
-    );
-    if (ignorePatterns && ignorePatterns.content) {
-      this.ignoreMap.set(url, ignore().add(ignorePatterns.content));
-    }
-
-    for await (let { path: innerPath, kind } of this.readdir(
-      this.realmPaths.local(url)
-    )) {
-      let innerURL = this.realmPaths.fileURL(innerPath);
-      if (this.isIgnored(innerURL)) {
-        continue;
-      }
-      if (kind === "file") {
-        await this.visitFile(innerURL);
-      } else {
-        let directoryURL = this.realmPaths.directoryURL(innerPath);
-        await this.visitDirectory(directoryURL);
-      }
-    }
+    this.#currentRun = await CurrentRun.fromScratch(this.realm, {
+      readdir: this.readdir,
+      readFileAsText: this.readFileAsText,
+    });
   }
 
   async update(url: URL, opts?: { delete?: true }): Promise<void> {
-    await this.visitFile(url, opts);
-    await this.semanticPhase();
-  }
-
-  private async visitFile(url: URL, opts?: { delete?: true }): Promise<void> {
-    if (this.isIgnored(url)) {
-      return;
-    }
-
-    let localPath = this.realmPaths.local(url);
-    let fileRef = await this.readFileAsText(localPath);
-    if (!fileRef) {
-      return;
-    }
-    let { content, lastModified } = fileRef;
-    if (url.href.endsWith(".json")) {
-      let json = JSON.parse(content);
-      if (isCardDocument(json)) {
-        let instanceURL = new URL(url.href.replace(/\.json$/, ""));
-        if (opts?.delete && this.instances.get(instanceURL)) {
-          this.instances.remove(instanceURL);
-        } else {
-          json.data.id = instanceURL.href;
-          json.data.meta.lastModified = lastModified;
-          let module = await Loader.import<Record<string, any>>(
-            new URL(
-              json.data.meta.adoptsFrom.module,
-              new URL(localPath, this.realm.url)
-            ).href
-          );
-          let CardClass = module[json.data.meta.adoptsFrom.name] as typeof Card;
-          let card = CardClass.fromSerialized(json.data.attributes);
-          let searchData = await this.api.searchDoc(card);
-          this.instances.set(instanceURL, {
-            resource: json.data,
-            searchData,
-            types: undefined,
-          });
-        }
-      }
-    } else if (
-      hasExecutableExtension(url.href) &&
-      url.href !== `${baseRealm.url}card-api.gts` // the base card's module is not analyzable
-    ) {
-      if (opts?.delete) {
-        if (this.modules.get(url)) {
-          this.modules.remove(url);
-        }
-        if (this.modules.get(trimExecutableExtension(url))) {
-          this.modules.remove(trimExecutableExtension(url));
-        }
-      } else {
-        let mod = new ModuleSyntax(content);
-        this.modules.set(url, mod);
-        this.modules.set(trimExecutableExtension(url), mod);
-      }
-    }
-  }
-
-  private async semanticPhase(): Promise<void> {
-    let newDefinitions: Map<string, CardDefinition> = new Map([
-      // seed the definitions with the base card
-      [
-        this.internalKeyFor(
-          {
-            type: "exportedCard",
-            module: `${baseRealm.url}card-api`,
-            name: "Card",
-          },
-          undefined
-        ),
-        {
-          id: {
-            type: "exportedCard",
-            module: `${baseRealm.url}card-api`,
-            name: "Card",
-          },
-          key: this.internalKeyFor(
-            {
-              type: "exportedCard",
-              module: `${baseRealm.url}card-api`,
-              name: "Card",
-            },
-            undefined
-          ),
-          super: undefined,
-          fields: new Map(),
-        },
-      ],
-    ]);
-    for (let [url, mod] of this.modules) {
-      for (let possibleCard of mod.possibleCards) {
-        if (possibleCard.exportedAs) {
-          if (this.isIgnored(url)) {
-            continue;
-          }
-          await this.buildDefinition(
-            newDefinitions,
-            url,
-            mod,
-            {
-              type: "exportedCard",
-              module: url.href,
-              name: possibleCard.exportedAs,
-            },
-            possibleCard
-          );
-        }
-      }
-    }
-    let newExportedCardRefs = new Map<string, Map<string, ExportedCardRef>>();
-    for (let def of newDefinitions.values()) {
-      if (def.id.type !== "exportedCard") {
-        continue;
-      }
-      let { module } = def.id;
-      let refsMap = newExportedCardRefs.get(module);
-      if (!refsMap) {
-        refsMap = new Map();
-        newExportedCardRefs.set(module, refsMap);
-      }
-      let { type: remove, ...exportedCardRef } = def.id;
-      refsMap.set(this.internalKeyFor(def.id, undefined), exportedCardRef);
-    }
-
-    // atomically update the search index
-    this.definitions = newDefinitions;
-    this.exportedCardRefs = newExportedCardRefs;
-
-    // once we have definitions we can fill in the instance types
-    for (let [url, entry] of [...this.instances]) {
-      entry.types = await this.getTypes(entry.resource.meta.adoptsFrom, url);
-    }
-  }
-
-  private async buildDefinition(
-    definitions: Map<string, CardDefinition>,
-    url: URL,
-    mod: ModuleSyntax,
-    ref: CardRef,
-    possibleCard: PossibleCardClass
-  ): Promise<CardDefinition | undefined> {
-    let id: CardRef = possibleCard.exportedAs
-      ? {
-          type: "exportedCard",
-          module: trimExecutableExtension(new URL(url, this.realm.url)).href,
-          name: possibleCard.exportedAs,
-        }
-      : ref;
-
-    let def = definitions.get(this.internalKeyFor(id, url));
-    if (def) {
-      definitions.set(this.internalKeyFor(ref, url), def);
-      return def;
-    }
-
-    let superDef = await this.definitionForClassRef(
-      definitions,
+    this.#currentRun = await CurrentRun.incremental(
       url,
-      mod,
-      possibleCard.super,
-      { type: "ancestorOf", card: id }
+      opts?.delete ? "delete" : "update",
+      this.#currentRun
     );
-
-    if (!superDef) {
-      return undefined;
-    }
-
-    let fields: CardDefinition["fields"] = new Map(superDef.fields);
-
-    for (let [fieldName, possibleField] of possibleCard.possibleFields) {
-      if (!isOurFieldDecorator(possibleField.decorator, url)) {
-        continue;
-      }
-      let fieldType = getFieldType(possibleField.type, url);
-      if (!fieldType) {
-        continue;
-      }
-      let fieldDef = await this.definitionForClassRef(
-        definitions,
-        url,
-        mod,
-        possibleField.card,
-        { type: "fieldOf", card: id, field: fieldName }
-      );
-      if (fieldDef) {
-        fields.set(fieldName, { fieldType, fieldCard: fieldDef.id });
-      }
-    }
-
-    let key = this.internalKeyFor(id, url);
-    def = { id, key, super: superDef.id, fields };
-    definitions.set(key, def);
-    return def;
-  }
-
-  private async definitionForClassRef(
-    definitions: Map<string, CardDefinition>,
-    url: URL,
-    mod: ModuleSyntax,
-    ref: ClassReference,
-    targetRef: CardRef
-  ): Promise<CardDefinition | undefined> {
-    if (ref.type === "internal") {
-      return await this.buildDefinition(
-        definitions,
-        url,
-        mod,
-        targetRef,
-        mod.possibleCards[ref.classIndex]
-      );
-    } else {
-      if (this.isLocal(new URL(ref.module, url))) {
-        if (
-          baseRealm.fileURL(ref.module).href === `${baseRealm.url}card-api` &&
-          ref.name === "Card"
-        ) {
-          let { module, name } = ref;
-          return definitions.get(
-            this.internalKeyFor({ module, name, type: "exportedCard" }, url)
-          );
-        }
-        let inner = this.lookupPossibleCard(new URL(ref.module, url), ref.name);
-        if (!inner) {
-          return undefined;
-        }
-        return await this.buildDefinition(
-          definitions,
-          new URL(ref.module, url),
-          inner.mod,
-          targetRef,
-          inner.possibleCard
-        );
-      } else {
-        return await this.getExternalCardDefinition({
-          name: ref.name,
-          module: ref.module,
-        });
-      }
-    }
-  }
-
-  private internalKeyFor(ref: CardRef, relativeTo: URL | undefined): string {
-    switch (ref.type) {
-      case "exportedCard":
-        let module = trimExecutableExtension(
-          new URL(ref.module, relativeTo)
-        ).href;
-        return `${module}/${ref.name}`;
-      case "ancestorOf":
-        return `${this.internalKeyFor(ref.card, relativeTo)}/ancestor`;
-      case "fieldOf":
-        return `${this.internalKeyFor(ref.card, relativeTo)}/fields/${
-          ref.field
-        }`;
-    }
-  }
-
-  private lookupPossibleCard(
-    module: URL,
-    exportedName: string
-  ): { mod: ModuleSyntax; possibleCard: PossibleCardClass } | undefined {
-    let mod = this.modules.get(module);
-    if (!mod) {
-      // TODO: broken import seems bad
-      return undefined;
-    }
-    let possibleCard = mod.possibleCards.find(
-      (c) => c.exportedAs === exportedName
-    );
-    if (!possibleCard) {
-      return undefined;
-    }
-    return { mod, possibleCard };
-  }
-
-  private isLocal(url: URL): boolean {
-    return url.href.startsWith(this.realm.url);
   }
 
   async search(query: Query): Promise<CardResource[]> {
@@ -554,31 +195,26 @@ export class SearchIndex {
       name: "Card",
     });
 
-    return [...this.instances.values()]
+    return [...this.#currentRun.instances.values()]
       .filter(matcher)
       .sort(this.buildSorter(query.sort))
       .map((entry) => entry.resource);
+  }
+
+  public isIgnored(url: URL): boolean {
+    return this.#currentRun.isIgnored(url);
   }
 
   async typeOf(
     ref: CardRef,
     relativeTo = new URL(this.realm.url)
   ): Promise<CardDefinition | undefined> {
-    let def = this.definitions.get(this.internalKeyFor(ref, relativeTo));
-    if (def) {
-      return def;
-    }
-    let { module } = getExportedCardContext(ref);
-    let moduleURL = new URL(module, relativeTo);
-    if (!this.realm.paths.inRealm(moduleURL) && ref.type === "exportedCard") {
-      return await this.getExternalCardDefinition(ref);
-    }
-    return undefined;
+    return await this.#currentRun.typeOf(ref, relativeTo);
   }
 
   async exportedCardsOf(module: string): Promise<ExportedCardRef[]> {
     module = trimExecutableExtension(new URL(module, this.realm.url)).href;
-    let refsMap = this.exportedCardRefs.get(module);
+    let refsMap = this.#currentRun.exportedCardRefs.get(module);
     if (!refsMap) {
       return [];
     }
@@ -586,47 +222,12 @@ export class SearchIndex {
   }
 
   async card(url: URL): Promise<CardResource | undefined> {
-    return this.instances.get(url)?.resource;
-  }
-
-  private async getTypes(
-    ref: ExportedCardRef,
-    relativeTo = new URL(ref.module)
-  ): Promise<string[]> {
-    let fullRef: CardRef | undefined = { type: "exportedCard", ...ref };
-    let types: string[] = [];
-    while (fullRef) {
-      let def: CardDefinition | undefined = await this.typeOf(
-        fullRef,
-        relativeTo
-      );
-      if (!def) {
-        // TODO: create a way to report this error without breaking indexing
-        throw new Error(
-          `Tried to getTypes of ${JSON.stringify(
-            ref
-          )} but couldn't find that definition`
-        );
-      }
-      types.push(this.internalKeyFor(fullRef, relativeTo));
-      fullRef = def.super;
-    }
-    return types;
+    return this.#currentRun.instances.get(url)?.resource;
   }
 
   private cardHasType(entry: SearchEntry, ref: CardRef): boolean {
     return Boolean(
-      entry.types?.find((t) => t === this.internalKeyFor(ref, undefined)) // assumes ref refers to absolute module URL
-    );
-  }
-
-  private async getCardDefinition(
-    ref: ExportedCardRef
-  ): Promise<CardDefinition | undefined> {
-    return (
-      this.definitions.get(
-        this.internalKeyFor({ type: "exportedCard", ...ref }, undefined) // assumes ref refers to absolute module URL
-      ) ?? (await this.getExternalCardDefinition(ref))
+      entry.types?.find((t) => t === internalKeyFor(ref, undefined)) // assumes ref refers to absolute module URL
     );
   }
 
@@ -634,7 +235,7 @@ export class SearchIndex {
     ref: ExportedCardRef,
     fieldSegments: string[]
   ): Promise<CardDefinition | undefined> {
-    let def = await this.getCardDefinition(ref);
+    let def = await this.#currentRun.getCardDefinition(ref);
     if (!def) {
       return undefined;
     }
@@ -642,7 +243,7 @@ export class SearchIndex {
     let fieldDef = def.fields.get(fieldName);
     if (!fieldDef) {
       throw new Error(
-        `Your filter refers to nonexistent field "${fieldName}" on type ${this.internalKeyFor(
+        `Your filter refers to nonexistent field "${fieldName}" on type ${internalKeyFor(
           { type: "exportedCard", ...ref },
           undefined // assumes absolute module URL
         )}`
@@ -660,7 +261,7 @@ export class SearchIndex {
         ...fieldSegments,
       ]);
     }
-    return this.getCardDefinition(fieldDef.fieldCard);
+    return this.#currentRun.getCardDefinition(fieldDef.fieldCard);
   }
 
   private async loadFieldCard(
@@ -800,7 +401,7 @@ export class SearchIndex {
       return (entry) =>
         every(Object.entries(filter.eq), ([fieldPath, value]) => {
           if (this.cardHasType(entry, ref)) {
-            let queryValue = this.api.getQueryableValue(
+            let queryValue = this.#currentRun.api.getQueryableValue(
               fieldCards[fieldPath]!,
               value
             );
@@ -843,22 +444,34 @@ export class SearchIndex {
               (range.gt &&
                 !(
                   value >
-                  this.api.getQueryableValue(fieldCards[fieldPath]!, range.gt)
+                  this.#currentRun.api.getQueryableValue(
+                    fieldCards[fieldPath]!,
+                    range.gt
+                  )
                 )) ||
               (range.lt &&
                 !(
                   value <
-                  this.api.getQueryableValue(fieldCards[fieldPath]!, range.lt)
+                  this.#currentRun.api.getQueryableValue(
+                    fieldCards[fieldPath]!,
+                    range.lt
+                  )
                 )) ||
               (range.gte &&
                 !(
                   value >=
-                  this.api.getQueryableValue(fieldCards[fieldPath]!, range.gte)
+                  this.#currentRun.api.getQueryableValue(
+                    fieldCards[fieldPath]!,
+                    range.gte
+                  )
                 )) ||
               (range.lte &&
                 !(
                   value <=
-                  this.api.getQueryableValue(fieldCards[fieldPath]!, range.lte)
+                  this.#currentRun.api.getQueryableValue(
+                    fieldCards[fieldPath]!,
+                    range.lte
+                  )
                 ))
             ) {
               return false;
@@ -884,97 +497,23 @@ export class SearchIndex {
     }
     return def;
   }
-
-  public isIgnored(url: URL): boolean {
-    if (url.href === this.realm.url) {
-      return false; // you can't ignore the entire realm
-    }
-    if (this.ignoreMap.size === 0) {
-      return false;
-    }
-    // Test URL against closest ignore. (Should the ignores cascade? so that the
-    // child ignore extends the parent ignore?)
-    let ignoreURLs = [...this.ignoreMap.keys()].map((u) => u.href);
-    let matchingIgnores = ignoreURLs.filter((u) => url.href.includes(u));
-    let ignoreURL = matchingIgnores.sort((a, b) => b.length - a.length)[0] as
-      | string
-      | undefined;
-    if (!ignoreURL) {
-      return false;
-    }
-    let ignore = this.ignoreMap.get(new URL(ignoreURL))!;
-    let pathname = this.realmPaths.local(url);
-    return ignore.test(pathname).ignored;
-  }
-
-  private async getExternalCardDefinition(
-    ref: ExportedCardRef
-  ): Promise<CardDefinition | undefined> {
-    let key = this.internalKeyFor({ type: "exportedCard", ...ref }, undefined); // these should always be absolute URLs
-    let promise = this.#externalDefinitionsCache.get(key);
-    if (promise) {
-      return await promise;
-    }
-    let deferred = new Deferred<CardDefinition | undefined>();
-    this.#externalDefinitionsCache.set(key, deferred.promise);
-
-    let url = `${ref.module}/_typeOf?${stringify({
-      type: "exportedCard",
-      ...ref,
-    })}`;
-    let response = await Loader.fetch(url, {
-      headers: {
-        Accept: "application/vnd.api+json",
-      },
-    });
-    if (!response.ok) {
-      console.log(`Could not get card type for ${url}: ${response.status}`);
-      deferred.fulfill(undefined);
-      return undefined;
-    }
-
-    let resource: CardDefinitionResource = (await response.json()).data;
-    let def: CardDefinition = {
-      id: resource.attributes.cardRef,
-      key: resource.id,
-      super: resource.relationships._super?.meta.ref,
-      fields: new Map(
-        Object.entries(resource.relationships)
-          .filter(([fieldName]) => fieldName !== "_super")
-          .map(([fieldName, fieldInfo]) => [
-            fieldName,
-            {
-              fieldType: fieldInfo.meta.type as "contains" | "containsMany",
-              fieldCard: fieldInfo.meta.ref,
-            },
-          ])
-      ),
-    };
-    deferred.fulfill(def);
-    return def;
-  }
 }
 
-function isOurFieldDecorator(ref: ClassReference, inModule: URL): boolean {
-  return (
-    ref.type === "external" &&
-    new URL(ref.module, inModule).href === baseRealm.fileURL("card-api").href &&
-    ref.name === "field"
-  );
-}
-
-function getFieldType(
-  ref: ClassReference,
-  inModule: URL
-): "contains" | "containsMany" | undefined {
-  if (
-    ref.type === "external" &&
-    new URL(ref.module, inModule).href === baseRealm.fileURL("card-api").href &&
-    ["contains", "containsMany"].includes(ref.name)
-  ) {
-    return ref.name as ReturnType<typeof getFieldType>;
+export function internalKeyFor(
+  ref: CardRef,
+  relativeTo: URL | undefined
+): string {
+  switch (ref.type) {
+    case "exportedCard":
+      let module = trimExecutableExtension(
+        new URL(ref.module, relativeTo)
+      ).href;
+      return `${module}/${ref.name}`;
+    case "ancestorOf":
+      return `${internalKeyFor(ref.card, relativeTo)}/ancestor`;
+    case "fieldOf":
+      return `${internalKeyFor(ref.card, relativeTo)}/fields/${ref.field}`;
   }
-  return undefined;
 }
 
 // three-valued version of Array.every that propagates nulls. Here, the presence
